pyproject.toml
[project]
name = "asx-session-alerts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
